requirements.txt
apache-airflow-providers-amazon>=7.3.0,<10
apache-airflow-providers-common-sql>=1.3.2,<2
apache-airflow-providers-postgres>=5.4.0,<7
apache-airflow-providers-slack>=9,<10
apache-airflow[pandas]==3.0.2
beautifulsoup4>=4,<5
requests>=2.28.0,<3
yfinance>=0.2.63,<0.3
parquet-tools>=0.2.13,<1
pendulum>=3,<4
